+++ src/component/modal/CommonModalStyle.js
