+++ src/page/Meeting.js
@@ -1,5 +1,5 @@
 import Header from "../component/header/Header";
-import SearchOption from "../data/SearchOption";
+import SearchOption from "../data/searchOption";
 
 function MeetingPage() {
   return <Header searchOption={SearchOption.meeting}/>
