+++ src/page/Member.js
@@ -1,5 +1,5 @@
 import Header from "../component/header/Header";
-import Status from "../data/SearchOption";
+import Status from "../data/searchOption";
 
 function MemberPage() {
   return <Header searchOption={Status.member}/>
