+++ src/page/Reservation.js
@@ -1,5 +1,5 @@
 import Header from "../component/header/Header";
-import SearchOption from "../data/SearchOption";
+import SearchOption from "../data/searchOption";
 
 function ReservationPage() {
   return <Header searchOption={SearchOption.reservation} />
